requirements.txt
fastapi
uvicorn
pydantic
aio_pika
python-dotenv
websockets
requests